server-python/requirements.txt
flask
rembg
pillow
numpy
scikit-image
scipy
onnxruntime
